LLM/requirements.txt
langchain
#langchain_chroma
langchain_core
langchain_text_splitters
langchain_openai
langchain_community
pypdf
flask
openai
python-dotenv 
psycopg2